services/text_extraction.py
# services/text_extraction.py
from fastapi import HTTPException
import pymupdf
import docx
from io import BytesIO

def extract_text_from_pdf(file_bytes: bytes) -> str:
    try:
        doc = pymupdf.open(stream=file_bytes)
        text = ""
        for page in doc:
            text += page.get_text()
        doc.close()
        return text
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Error extracting PDF: {str(e)}")

def extract_text_from_docx(file_bytes: bytes) -> str:
    try:
        doc = docx.Document(BytesIO(file_bytes))
        text = ""
        for paragraph in doc.paragraphs:
            text += paragraph.text + "\n"
        return text
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Error extracting DOCX: {str(e)}")

